Separate orchestrator and service classes from handlers
